requirements.txt
discord.py
fastapi
uvicorn
python-dotenv
aiosqlite